generate all vocab types from schemas (#257)
