+++ src/lib/vocab/account/account.ts
@@ -1,3 +1,11 @@
+// generated by src/lib/vocab/account/account.schema.ts
+
+export interface AccountModel {
+	account_id: number;
+	name: string;
+	created: Date;
+	updated: Date | null;
+}
 export interface Account {
 	account_id: number;
 	name: string;
@@ -5,43 +13,5 @@ export interface Account {
 	created: Date;
 	updated: Date | null;
 }
-export const AccountSchema = {
-	$id: '/schemas/Account.json',
-	type: 'object',
-	properties: {
-		account_id: {type: 'number'},
-		name: {type: 'string'},
-		password: {type: 'string'},
-		created: {type: 'object', format: 'date-time', tsType: 'Date'},
-		updated: {type: ['object', 'null'], format: 'date-time', tsType: 'Date | null'},
-	},
-	required: ['account_id', 'name', 'password', 'created', 'updated'],
-	additionalProperties: false,
-};
 
-// TODO make the `create_account` event in account.events.ts
-export interface CreateAccountParams {
-	name: string;
-	password: string;
-}
-
-// TODO rename? `AccountClientDoc`? above could be `AccountDbDoc` and `AccountRequestDoc`
-export interface AccountModel {
-	account_id: number;
-	name: string;
-	created: Date;
-	updated: Date | null;
-}
-
-export const AccountModelSchema = {
-	$id: '/schemas/AccountModel.json',
-	type: 'object',
-	properties: {
-		account_id: {type: 'number'},
-		name: {type: 'string'},
-		created: {type: 'object', format: 'date-time', tsType: 'Date'},
-		updated: {type: ['object', 'null'], format: 'date-time', tsType: 'Date | null'},
-	},
-	required: ['account_id', 'name', 'created', 'updated'],
-	additionalProperties: false,
-};
+// generated by src/lib/vocab/account/account.schema.ts
