+++ src/lib/vocab/account/validateAccount.ts
@@ -1,10 +1,6 @@
 import {toValidateSchema} from '$lib/util/ajv';
-import {
-	AccountSchema,
-	type Account,
-	type AccountModel,
-	AccountModelSchema,
-} from '$lib/vocab/account/account';
+import {type Account, type AccountModel} from '$lib/vocab/account/account';
+import {AccountSchema, AccountModelSchema} from '$lib/vocab/account/account.schema';
 
 export const validateAccount = toValidateSchema<Account>(AccountSchema);
 export const validateAccountModel = toValidateSchema<AccountModel>(AccountModelSchema);
