+++ src/lib/vocab/community/validateCommunity.ts
@@ -1,5 +1,5 @@
 import {toValidateSchema} from '$lib/util/ajv';
-import {CommunitySchema} from '$lib/vocab/community/community';
-import type {Community} from '$lib/vocab/community/community';
+import {CommunitySchema} from '$lib/vocab/community/community.schema';
+import {type Community} from '$lib/vocab/community/community';
 
 export const validateCommunity = toValidateSchema<Community>(CommunitySchema);
