+++ src/lib/vocab/entity/entity.ts
@@ -1,5 +1,6 @@
-import {GUEST_PERSONA_NAME} from '$lib/vocab/persona/constants';
-import type {EntityData} from '$lib/vocab/entity/entityData';
+// generated by src/lib/vocab/entity/entity.schema.ts
+
+import {type EntityData} from '$lib/vocab/entity/entityData';
 
 export interface Entity {
 	entity_id: number;
@@ -9,26 +10,5 @@ export interface Entity {
 	created: Date;
 	updated: Date | null;
 }
-export const EntitySchema = {
-	$id: '/schemas/Entity.json',
-	type: 'object',
-	properties: {
-		entity_id: {type: 'number'},
-		actor_id: {type: 'number'},
-		space_id: {type: 'number'},
-		data: {type: 'object', tsType: 'EntityData'},
-		created: {type: 'object', format: 'date-time', tsType: 'Date'},
-		updated: {type: ['object', 'null'], format: 'date-time', tsType: 'Date | null'},
-	},
-	required: ['entity_id', 'actor_id', 'space_id', 'data', 'created', 'updated'],
-	additionalProperties: false,
-};
 
-// TODO expand to the entire vocabulary? generate if so
-export type EntityType = 'Persona' | 'Community';
-
-export const toName = (entity: null | undefined | {name?: string}): string =>
-	(entity as any)?.name ?? GUEST_PERSONA_NAME;
-
-export const toIcon = (entity: null | undefined | {icon?: string}): string | null =>
-	(entity as any)?.icon ?? null;
+// generated by src/lib/vocab/entity/entity.schema.ts
