+++ src/lib/vocab/membership/membership.schema.ts
@@ -1,0 +1,12 @@
+export const MembershipSchema = {
+	$id: '/schemas/Membership.json',
+	type: 'object',
+	properties: {
+		persona_id: {type: 'number'},
+		community_id: {type: 'number'},
+		created: {type: 'object', format: 'date-time', tsType: 'Date'},
+		updated: {type: ['object', 'null'], format: 'date-time', tsType: 'Date | null'},
+	},
+	required: ['persona_id', 'community_id', 'created', 'updated'],
+	additionalProperties: false,
+};
