+++ src/lib/vocab/persona/persona.ts
@@ -1,3 +1,8 @@
+// generated by src/lib/vocab/persona/persona.schema.ts
+
+export interface PersonaParams {
+	name: string;
+}
 export interface Persona {
 	persona_id: number;
 	account_id?: number;
@@ -8,35 +13,5 @@ export interface Persona {
 	created: Date;
 	updated: Date | null;
 }
-export const PersonaSchema = {
-	$id: '/schemas/Persona.json',
-	type: 'object',
-	properties: {
-		persona_id: {type: 'number'},
-		account_id: {type: 'number'},
-		community_id: {type: 'number'},
-		type: {type: 'string', enum: ['account', 'community']},
-		name: {type: 'string'},
-		icon: {type: 'string'},
-		created: {type: 'object', format: 'date-time', tsType: 'Date'},
-		updated: {type: ['object', 'null'], format: 'date-time', tsType: 'Date | null'},
-	},
-	required: ['persona_id', 'community_id', 'type', 'name', 'created', 'updated'],
-	additionalProperties: false,
-};
 
-export interface PersonaParams {
-	name: string;
-}
-export const PersonaParamsSchema = {
-	$id: '/schemas/PersonaParams.json',
-	type: 'object',
-	properties: {
-		name: {type: 'string'},
-	},
-	required: ['name'],
-	additionalProperties: false,
-};
-
-//TODO
-//2.5: Render active persona
+// generated by src/lib/vocab/persona/persona.schema.ts
