+++ src/lib/vocab/persona/validatePersona.ts
@@ -1,6 +1,6 @@
 import {toValidateSchema} from '$lib/util/ajv';
-import {PersonaParamsSchema, PersonaSchema} from '$lib/vocab/persona/persona';
-import type {Persona, PersonaParams} from '$lib/vocab/persona/persona';
+import {PersonaParamsSchema, PersonaSchema} from '$lib/vocab/persona/persona.schema';
+import {type Persona, type PersonaParams} from '$lib/vocab/persona/persona';
 
 export const validatePersona = toValidateSchema<Persona>(PersonaSchema);
